+++ wp-content/themes/twentytwenty-child/components/product-data.php
@@ -17,10 +17,10 @@
         <?php
 
         $isOnSale = get_post_meta( $post->ID, 'products_fields_is_on_sale', true );
-        $price = number_format( (int) get_post_meta( $post->ID, 'products_fields_price', true ),2);
+        $price = number_format( (float) get_post_meta( $post->ID, 'products_fields_price', true ),2);
 
         if (get_post_meta( $post->ID, 'products_fields_sale_price', true )){
-            $salePrice = number_format( (int) get_post_meta( $post->ID, 'products_fields_sale_price', true ),2);
+            $salePrice = number_format( (float) get_post_meta( $post->ID, 'products_fields_sale_price', true ),2);
         }else{
             $salePrice = false;
         }
